Refactor game-core into ai and ui modules
