+++ Progress.md
@@ -3,7 +3,7 @@
 - [x] **8方向移動 & マップスクロール** – `@` を矢印/viキー対応で移動。境界外には出られない。
 - [x] **ターン制ゲームループ** – 入力 → 状態更新 → 描画 の順で一括処理。
 - [x] **キー操作時のみターン進行** – 無入力では時間が止まり、クラシックローグライクのテンポに。
-- [ ] **コードベース整理 (`src/ai`, `src/ui` 等)** – 今後の追加機能に合わせてサブモジュールを分離し見通しを確保。
+- [x] **コードベース整理 (`src/ai`, `src/ui` 等)** – 今後の追加機能に合わせてサブモジュールを分離し見通しを確保。
 
 ## 1. 探索 & 生態系
 - [x] **魚スポーン & ランダム移動AI** – 深度レンジを持つ魚を水域に配置し、毎ターン微移動。
