+++ docs/designs.md
@@ -26,6 +26,7 @@
 * **ui**: 描画ラッパ & ウィジェット。
 * **assets/**: JSON データ + RON セーブファイル。
 * **common**: 共有の型とエラー定義。
+* **game-core 内部モジュール**: `ai.rs` と `ui.rs` にロジックと描画処理を分割し拡張に備える。
 
 ## 2. 実行環境 / ビルドターゲット
